+++ .env_example
@@ -1,2 +1,12 @@
-TOKEN=TEST
-PROVIDER_TOKEN=TEST
+TOKEN=7531152085:AAHmU9rQzRSQKtHthGASHgu29p5l3anc7lE
+PROVIDER_TOKEN=381764678:TEST:127140
+
+DB_USER=username
+DB_PASSWORD=password
+DB_NAME=bot_database
+
+API_ID=25515264
+API_HASH=da8436b7652b1147590f7307d7420f02
+PHONE=+79101558161
+
+GROUP_ID=0
