+++ .gitignore
@@ -2,3 +2,4 @@
 venv
 .env
 config.py
+*.session*
